+++ src/main/java/cart/auth/AuthInterceptor.java
@@ -1,5 +1,7 @@
 package cart.auth;
 
+import cart.dao.CredentialDao;
+import cart.exception.AuthenticationException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import org.springframework.web.servlet.HandlerInterceptor;
