+++ src/main/java/cart/config/WebConfiguration.java
@@ -3,8 +3,8 @@ package cart.config;
 import cart.auth.AuthArgumentResolver;
 import cart.auth.AuthInterceptor;
 import cart.auth.BasicAuthorizationParser;
-import cart.auth.CredentialDao;
 import cart.auth.CredentialThreadLocal;
+import cart.dao.CredentialDao;
 import java.util.List;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
