+++ src/main/java/cart/exception/AuthenticationException.java
@@ -1,4 +1,4 @@
-package cart.auth;
+package cart.exception;
 
 public class AuthenticationException extends RuntimeException {
