+++ src/main/java/cart/exception/InvalidBasicCredentialException.java
@@ -1,4 +1,4 @@
-package cart.auth;
+package cart.exception;
 
 public class InvalidBasicCredentialException extends RuntimeException {
